# docs/swagger.yaml
basePath: /api
definitions:
  entity.AuthResponse:
    properties:
      expires_at:
        type: string
      refresh_token:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/entity.MemberResponse'
    type: object
  entity.LoginRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 20
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  entity.MemberResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      status:
        $ref: '#/definitions/entity.MemberStatus'
      updated_at:
        type: string
    type: object
  entity.MemberStatus:
    enum:
    - ACTIVE
    - INACTIVE
    - SUSPENDED
    - REJECTED
    type: string
    x-enum-varnames:
    - MemberStatusActive
    - MemberStatusInactive
    - MemberStatusSuspended
    - MemberStatusRejected
  entity.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        maxLength: 20
        minLength: 6
        type: string
      phone:
        type: string
    required:
    - email
    - name
    - password
    - phone
    type: object
info:
  contact:
    name: Milkly Team
  description: API for Milkly Member Management
  title: Milkly Member Service API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a member and return a JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Login member
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a member and return a JWT token
      parameters:
      - description: Registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Register member
      tags:
      - auth
  /health:
    get:
      consumes:
      - application/json
      description: Check if the service is running
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Health check endpoint
      tags:
      - health
swagger: "2.0"